refactor(workspace): eliminate feature flag duplication
Remove 27 optional deps from root Cargo.toml — each now lives only in
the crate that uses it. Simplify feature forwarding to single-target:

  Before: channel-nostr = ["agent-runtime", "zeroclaw-misc/channel-nostr",
           "zeroclaw-config/channel-nostr", "dep:nostr-sdk"]
  After:  channel-nostr = ["zeroclaw-misc/channel-nostr"]

Root features forward to ONE crate. That crate handles internal
forwarding (misc → channels → config). No root-level dep: for things
root doesn't use.

Fix otel regression test to check misc Cargo.toml instead of root.
Fix clippy: collapsible_if, redundant guard, prefix stripping,
empty lines, duplicated attributes.

All gates pass:
  cargo fmt: PASS
  cargo clippy --all-targets --all-features -D warnings: PASS
  cargo check (default, no-default, all-features): PASS
  cargo test --workspace --all-features: 6,918 passed, 0 failed

+++ crates/zeroclaw-misc/src/observability/otel.rs
@@ -445,13 +445,13 @@ impl Observer for OtelObserver {
                 self.request_latency.record(d.as_secs_f64(), &[]);
             }
             ObserverMetric::TokensUsed(t) => {
-                self.tokens_used.add(*t as u64, &[]);
+                self.tokens_used.add(*t, &[]);
             }
             ObserverMetric::ActiveSessions(s) => {
-                self.active_sessions.record(*s as u64, &[]);
+                self.active_sessions.record(*s, &[]);
             }
             ObserverMetric::QueueDepth(d) => {
-                self.queue_depth.record(*d as u64, &[]);
+                self.queue_depth.record(*d, &[]);
             }
             ObserverMetric::HandRunDuration {
                 hand_name,
